% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_forest.R
\name{rotationforest_train}
\alias{rotationforest_train}
\title{Internal function wrappers}
\usage{
rotationforest_train(
  x,
  y,
  npredictor,
  ntree,
  verbose,
  cp = 0.01,
  minsplit = 20L,
  maxdepth = 30L
)
}
\description{
These are not supported when called by the user.
}
\keyword{internal}
